moved steps & features to root dir
and add attachments to allure report
