+++ features/tutorial.feature
@@ -1,5 +1,6 @@
 Feature: showing off behave
 
+  @google
   Scenario: run a simple test
      Given we have behave installed
       When we implement a test
